Push latest features to master (#1)
* Fix JSON

* Fix image preview

* Make gifs smaller

* Make logo smaller

* Shrink images

* Adjust images

* Add links

* Add import processing

* Add generate_string

* Add get_text_editor, bytes_to_string and sp_call

* Add docstrings

* Add location path info and remove add, remove, exists from Entry

* Make Database more python-native

* Testing for typer

* Modify docstrings and options

* Fix reversed options and arguments

* Add verbose arguments

* Add log functions

* Add exists

* Add excepthook

* Add post mortem info

* Remove unnecessary database variable

* Add find

* Enable verbose output using argument instead of env var

* Fix switched bool

* Add _get_database

* Add __rich__

* Add input override

* Add add_entry

* Add __version__.py

* Remove unused import

* Add versioin

* Remove unused import

* Remove unused code

* Fix show_version

* Make show_version private

* Add entry removal

* Fix exception

* Fix entry add

* Add __rich__

* Remove _gen_entry_tables and add permission check

* Add emphasis

* Add levels test

* Add docstring

* Make levels bold

* Fix formatting

* Remove unused imports

* Rename clityper.py to cli.py

* Rename database.py to core.py

Add docstrings
Use tabulate_mapping to decrease boilerplate code

* Rename database.py to core.py

* Remove unused stuff

* Remove generate_string

* Add exception shortcuts

* Make exceptions more specific

* Add delete method for Database

* Add python native shortcut

* Make exception follow name format

* Add and update docstrings

* Update database option evaluation

* Add license notice for version string

* Remove __delete__ as method was already implemented

* Add docstrings

* Remove copyright notices and update type hints

* Remove unused dependencies and add new ones

* Update required python version

* Update Pipfile.lock

* Fix leading seperator bug

* Make to_dict private

* Revert changes

* Fix datatype bug

* Add initial test

* Add entry equality test

* Add context manager for testing

* Fix add and remove returning None instead of modified database

* Finish core.py testing

* Rename core.py to test_core.py

* Add docstrings and modify test

* Move tests

* Update python required for README.md

* Update Built With

* Add more gifs

+++ whereis/core.py
@@ -1,24 +1,12 @@
-# where-is: Finds config files.
-# Copyright (C) 2020 ALinuxPerson
-#
-# This program is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+"""The core of where-is. This is where the CLI frontend gets its objects from."""
 import json
 from typing import List, Dict, Union
 from pathlib import Path
 import os
 from whereis import exceptions, utils
 import shutil
+from rich.table import Table
+from rich.tabulate import tabulate_mapping
 
 
 class Entry:
@@ -31,7 +19,6 @@ class Entry:
         """
         self._name = name
         self._locations = locations
-        self.database: Database = Database()
 
     @property
     def name(self) -> str:
@@ -49,13 +36,15 @@ class Entry:
         Notes:
             The paths can be formatted.
             Here is the following format map:
-                {HOME}: <your home folder>
+                {HOME}: Your home folder.
+                {WHEREIS_CONFIG}: The where-is configuration folder.
+                {CONFIG_FOLDER}: The configuration folder.
 
         Returns:
             All of the locations an entry has.
         """
         return [
-            Path(self._format_path(os.path.join(os.path.sep, *location)))
+            Path(self._format_path(Path(os.path.join(os.path.sep, *location))))
             for location in self._locations
         ]
 
@@ -68,7 +57,7 @@ class Entry:
         """
         return {
             "name": self.name,
-            "locations": self._locations,  # type: ignore
+            "locations": list(self._locations),
         }
 
     @property
@@ -81,7 +70,7 @@ class Entry:
         return json.dumps(self.to_dict)
 
     @staticmethod
-    def _format_path(path: str) -> str:
+    def _format_path(path: Path) -> Path:
         """Formats a path.
 
         Args:
@@ -90,39 +79,24 @@ class Entry:
         Returns:
             The formatted path.
         """
-        format_map: Dict[str, str] = {
-            "HOME": str(Path().home()),
-            "WHEREIS_CONFIG": str(utils.config_folder()),
-            "CONFIG_FOLDER": str(utils.config_folder() / ".."),
+        format_map: Dict[str, Path] = {
+            "HOME": Path().home(),
+            "WHEREIS_CONFIG": utils.config_folder(),
+            "CONFIG_FOLDER": utils.config_folder().parent,
         }
         try:
-            return path.format(**format_map)
+            path_parts: List[str] = list(path.parts)
+            for index, part in enumerate(path_parts):
+                for name, value in format_map.items():
+                    if name in part:
+                        del path_parts[index]
+                        path_parts[index:index] = list(value.parts)
+
+            return Path(*path_parts)
         except (KeyError, IndexError, ValueError):
-            raise KeyError(f"Format map not supported for path '{path}'.") from None
-
-    def add(self) -> None:
-        """Adds the entry object to the database.
-
-        Returns:
-            Nothing.
-        """
-        return self.database.add(self)
-
-    def remove(self) -> None:
-        """Removes the entry object from the database.
-
-        Returns:
-            Nothing.
-        """
-        return self.database.remove(self)
-
-    def exists(self) -> bool:
-        """Checks if the entry object exists in the database.
-
-        Returns:
-            True if the entry object exists, else False.
-        """
-        return self in self.database.entries
+            raise exceptions.FormatMapError(
+                f"Format map not supported for path '{path}'."
+            ) from None
 
     def locations_exists(self) -> Dict[Path, bool]:
         """Does each location exist?
@@ -138,6 +112,38 @@ class Entry:
         except AttributeError:
             return False
 
+    def __rich__(self) -> Table:
+        """A shortcut to generate a table to generate configuration files found.
+
+        Returns:
+            A table object, usable by rich print instances.
+        """
+        columns: List[str] = ["Locations", "Exists", "Is File", "Is Folder"]
+        table: Table = Table(title="[bold purple]Config files found")
+        for column in columns:
+            table.add_column(column)
+        for location, exists in self.locations_exists().items():
+            formatted_location: str = f"[magenta]{location}"
+            formatted_exists: str = f"[red]{exists}" if not exists else f"[green4]{exists}"
+
+            if exists and location.is_file():
+                is_file: str = "[green4]True"
+            elif exists and not location.is_file():
+                is_file = "[red]False"
+            else:
+                is_file = "[red italic]Unknown"
+
+            if exists and location.is_dir():
+                is_dir: str = "[green]True"
+            elif exists and not location.is_dir():
+                is_dir = "[red]False"
+            else:
+                is_dir = "[red italic]Unknown"
+
+            table.add_row(formatted_location, formatted_exists, is_file, is_dir)
+
+        return table
+
     def __repr__(self) -> str:
         return f"<{self.__class__.__name__} object: name='{self.name}' locations={self.locations}>"
 
@@ -150,7 +156,6 @@ class Database:
             location: The location where the database is. Defaults to the config folder.
         """
         self._location = location
-        self._not_exists_ok: bool = False
 
     @property
     def location(self) -> Path:
@@ -162,11 +167,6 @@ class Database:
         Raises:
             NotADirectoryError: When the database folder doesn't exist.
         """
-        if (
-            not self._location.is_dir() or not self._location.exists()
-        ) and not self._not_exists_ok:
-            # keyword being 'folder'
-            raise NotADirectoryError("The database folder must exist!")
         return self._location
 
     @property
@@ -175,6 +175,9 @@ class Database:
 
         Returns:
             A list of dictionaries if the file owning that dictionary's suffix is '.json'.
+
+        Raises:
+            EntryParseError: If the entry JSON can't be decoded.
         """
         ret: List[Dict[str, Union[str, List[List[str]]]]] = []
         for entry in self.location.iterdir():
@@ -182,7 +185,7 @@ class Database:
                 try:
                     ret.append(json.loads(entry.read_text()))
                 except json.decoder.JSONDecodeError as error:
-                    raise ValueError(
+                    raise exceptions.EntryParseError(
                         f"Error parsing '{entry.absolute()}': {error}"
                     ) from None
 
@@ -199,12 +202,14 @@ class Database:
             An entry object.
 
         Raises:
-            KeyError: If the raw entry doesn't follow the json schema.
+            EntryParseError: If the raw entry doesn't follow the json schema.
         """
         try:
             return Entry(raw_entry["name"], *raw_entry["locations"])  # type: ignore
         except KeyError:
-            raise ValueError(f"JSON schema incorrect: {raw_entry}") from None
+            raise exceptions.EntryParseError(
+                f"JSON schema incorrect: {raw_entry}"
+            ) from None
 
     @property
     def entries(self) -> List[Entry]:
@@ -245,7 +250,7 @@ class Database:
             EntryDoesNotExistError: If the entry object doesn't exist in the database entries.
         """
         if entry not in self.entries:
-            raise exceptions.EntryDoesNotExistError("The database entry must exist.")
+            raise exceptions.EntryNotFoundError("The database entry must exist.")
         entry_to_delete: Path = self.location / f"{entry.name}.json"
         entry_to_delete.unlink()
 
@@ -254,12 +259,102 @@ class Database:
 
         Returns:
             Nothing.
+
+        Raises:
+            DatabaseExistsError: If the database exists.
         """
+        if self.exists():
+            raise exceptions.DatabaseExistsError("The database already exists!")
         sample_db: Path = Path(__file__).parent / "database"
-        self._not_exists_ok = True
         Path(self.location).mkdir()
         for path in sample_db.iterdir():
             shutil.copyfile(str(path), str(self.location / path.name))
+
+    def delete(self) -> None:
+        """Deletes the database if it exists.
+
+        Notes:
+            There is no way to implement a python native shortcut for this because __del__ gets called after program
+            exit, which means the database will get deleted after program exit.
+
+        Returns:
+            Nothing.
+
+        Raises:
+            DatabaseNotFoundError: If the database doesn't exist.
+        """
+        if not self.location.exists():
+            raise exceptions.DatabaseNotFoundError("The database doesn't exist!")
+        return shutil.rmtree(self.location)
+
+    def exists(self) -> bool:
+        """Check if the database exists.
+
+        Returns:
+            True if the folder contacting the database exists, else False
+        """
+        return self.location.exists()
+
+    def __rich__(self) -> Table:
+        """A shortcut to generate database info.
+
+        Returns:
+            A table, usable by rich print instances.
+        """
+        map_: Dict[str, Union[Path, List[Entry], bool]] = {
+            "Location": self.location,
+            "Entries": self.entries,
+            "Exists": self.exists(),
+        }
+        table: Table = tabulate_mapping(map_, title="Database Info")
+
+        return table
+
+    def __add__(self, other: Entry):
+        """A python native shortcut of database.add().
+
+        Args:
+            other: The other class (should be an Entry object)
+
+        Returns:
+            Database: The modified database.
+        """
+        self.add(other)
+        return self
+
+    def __sub__(self, other: Entry):
+        """A python native shortcut of database.remove().
+
+        Args:
+            other: The other class (should be an Entry object)
+
+        Returns:
+            Database: The modified database.
+        """
+        self.remove(other)
+        return self
+
+    def __enter__(self):
+        """Create a temporary database. Useful for testing.
+
+        Returns:
+            Database: A database object.
+        """
+        self.create()
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb) -> None:
+        """Deletes a database object after context manager exit. Useful for testing.
+
+        Args:
+            exc_type: The exception type.
+            exc_val: The exception value.
+            exc_tb: The traceback.
+
+        Returns:
+            Nothing.
+        """
+        self.delete()
 
     def __repr__(self) -> str:
         return f"<{self.__class__.__name__} object: location={self.location}>"
